Add flash update edit link in Navabar
- Add validation when country is empty
- Fix typing issues
- Move utils to new file
- Change actions styling on form
