+++ src/views/FlashUpdateForm/common.ts
@@ -1,0 +1,71 @@
+import {
+    analyzeErrors,
+    getErrorObject,
+    type Error,
+} from '@togglecorp/toggle-form';
+import {
+    type FormType,
+} from './schema';
+
+export type TabKeys = 'context' | 'actions' | 'focal';
+
+export function getNextStep(current: TabKeys, direction: 1 | -1) {
+    if (direction === 1) {
+        const mapping: { [key in TabKeys]?: TabKeys } = {
+            context: 'actions',
+            actions: 'focal',
+        };
+        return mapping[current];
+    }
+    const mapping: { [key in TabKeys]?: TabKeys } = {
+        focal: 'actions',
+        actions: 'context',
+    };
+    return mapping[current];
+}
+
+const fieldsInContext: (keyof FormType)[] = [
+    'country_district',
+    'references',
+    'hazard_type',
+    'title',
+    'situational_overview',
+    'graphics_files',
+    'map_files',
+];
+
+const fieldsInActions: (keyof FormType)[] = [
+    'actions_taken',
+];
+
+const fieldsInFocalPoints: (keyof FormType)[] = [
+    'originator_name',
+    'originator_title',
+    'originator_email',
+    'originator_phone',
+    'ifrc_name',
+    'ifrc_title',
+    'ifrc_email',
+    'ifrc_phone',
+];
+
+const tabToFieldsMap: Record<TabKeys, (keyof FormType)[]> = {
+    context: fieldsInContext,
+    actions: fieldsInActions,
+    focal: fieldsInFocalPoints,
+};
+
+export function checkTabErrors(error: Error<FormType> | undefined, tabKey: TabKeys) {
+    if (!analyzeErrors(error)) {
+        return false;
+    }
+
+    const fields = tabToFieldsMap[tabKey];
+    const fieldErrors = getErrorObject(error);
+
+    const hasErrorOnAnyField = fields.some(
+        (field) => analyzeErrors(fieldErrors?.[field]),
+    );
+
+    return hasErrorOnAnyField;
+}
