+++ src/views/FlashUpdateForm/styles.module.css
@@ -9,11 +9,15 @@
         flex-direction: column;
         gap: var(--go-ui-spacing-4xl);
 
-        .page-actions {
+        .actions {
             display: flex;
-            flex-grow: 1;
-            justify-content: center;
-            gap: var(--go-ui-spacing-lg);
+
+            .page-actions {
+                display: flex;
+                flex-grow: 1;
+                justify-content: center;
+                gap: var(--go-ui-spacing-lg);
+            }
         }
     }
 }
